Revert to full FP16 precision - hybrid caused OOM
- Hybrid precision (FP32 model + FP16 frames) used 6.77GB + needed 5.91GB more
- Exceeds 8GB VRAM on RTX 2060, causing CUDA out of memory
- Full FP16 uses ~4-5GB total, fits comfortably in 8GB
- RTX series handles FP16 well with good quality
- Reverts model.half() to be included again

+++ faster-propainter-main/watermark.py
@@ -360,8 +360,7 @@ def pipeline(
             )
             gt_flows_bi = (gt_flows_bi[0].half(), gt_flows_bi[1].half())
             fix_flow_complete = fix_flow_complete.half()
-            # SKIP model.half() - keep inpainting model in FP32 for quality on GTX GPUs
-            # model = model.half()
+            model = model.half()
 
         # ---- complete flow ----
         flow_length = gt_flows_bi[0].size(1)
